rename: slice이름 수정
