+++ src/components/ExtendSideBar/PostForm/PostForm.js
@@ -13,7 +13,7 @@ import { useEffect, useState } from "react";
 import { createPortal } from "react-dom";
 import { useDispatch, useSelector } from "react-redux";
 import { useParams } from "react-router";
-import { setPostPosition } from "redux/modules/centerSlice";
+import { setPostPosition } from "redux/modules/mapSlice";
 import { db, storage } from "server/config";
 import { FlexCenter, FlexColumn } from "styles/mixins";
 import * as Styled from "./PostForm.styles";
@@ -194,7 +194,6 @@ export const PostForm = ({ closePost, openshow }) => {
       postMutate();
     }
   };
-
   return (
     <Styled.PostFormBlock>
       <Styled.ExtendSidebar>
